+++ src/xUtil.cpp
@@ -422,7 +422,7 @@ long long setime(void)
 
 /* Convert an amount of bytes into a human readable string in the form
  * of 100B, 2G, 100M, 4K, and so forth. */
-void bytesToHuman(char *s, unsigned long long n)
+void bytesToHuman(char *s, uint64_t n)
 {
 	double d;
